+++ hicexplorer/reduceMatrix.py
@@ -41,9 +41,9 @@ def reduce_matrix(matrix, bins_to_merge, use_triu=True, diagonal=False):
     A sparse matrix.
 
     >>> from scipy.sparse import csr_matrix
-    >>> A = csr_matrix(np.array([[1,0],[0,1]]))
+    >>> A = csr_matrix(np.array([[1,0],[0,1]]), dtype=np.int32)
     >>> reduce_matrix(A, [(0,1)], diagonal=True).todense()
-    matrix([[2]])
+    matrix([[2]], dtype=int32)
     >>> A = csr_matrix(np.array([[5,5,2,2,0],[0,5,2,2,1],
     ... [0,0,1,1,0], [0,0,0,1,0], [0,0,0,0,0]]), dtype=np.int32)
     >>> A.todense()
@@ -51,23 +51,23 @@ def reduce_matrix(matrix, bins_to_merge, use_triu=True, diagonal=False):
             [0, 5, 2, 2, 1],
             [0, 0, 1, 1, 0],
             [0, 0, 0, 1, 0],
-            [0, 0, 0, 0, 0]], dtype=np.int32)
+            [0, 0, 0, 0, 0]], dtype=int32)
     >>> ll = [(0,1), (2,3), (4,)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[15,  8,  1],
             [ 0,  3,  0],
-            [ 0,  0,  0]], dtype=np.int32)
+            [ 0,  0,  0]], dtype=int32)
 
     >>> ll = [(0,1,2), (3,4)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[20,  6],
-            [ 0,  1]], dtype=np.int32)
+            [ 0,  1]], dtype=int32)
 
     >>> ll = [(0,1), (2,3), (4,)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=True).todense()
     matrix([[15,  8,  1],
             [ 8,  3,  0],
-            [ 1,  0,  0]], dtype=np.int32)
+            [ 1,  0,  0]], dtype=int32)
 
     Test symmetric matrix
     >>> A = csr_matrix(np.array([[2,2,1,1,1,1],[2,2,1,1,1,1],
@@ -78,65 +78,65 @@ def reduce_matrix(matrix, bins_to_merge, use_triu=True, diagonal=False):
             [1, 1, 1, 1, 1, 1],
             [1, 1, 1, 1, 1, 1],
             [1, 1, 1, 1, 1, 1],
-            [1, 1, 1, 1, 1, 1]], dtype=np.int32)
+            [1, 1, 1, 1, 1, 1]], dtype=int32)
     >>> ll = [(0,1), (2,3), (4,5)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[8, 4, 4],
             [4, 4, 4],
-            [4, 4, 4]], dtype=np.int32)
+            [4, 4, 4]], dtype=int32)
 
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=True).todense()
     matrix([[6, 4, 4],
             [4, 3, 4],
-            [4, 4, 3]], dtype=np.int32)
+            [4, 4, 3]], dtype=int32)
 
     >>> ll = [(0,1,2), (3,4,5)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[13,  9],
-            [ 9,  9]], dtype=np.int32)
+            [ 9,  9]], dtype=int32)
 
     Test symmetric matrix non consecutive
     >>> A = csr_matrix(np.array([[5,1,5,1,0],[0,1,2,1,1],
     ... [0,0,5,1,0], [0,0,0,1,0], [0,0,0,0,0]]), dtype=np.int32)
     >>> dia = dia_matrix(([A.diagonal()], [0]), shape=A.shape)
-    >>> A= csr_matrix(A + A.T - dia)
+    >>> A= csr_matrix(A + A.T - dia, dtype=np.int32)
     >>> A.todense()
     matrix([[5, 1, 5, 1, 0],
             [1, 1, 2, 1, 1],
             [5, 2, 5, 1, 0],
             [1, 1, 1, 1, 0],
-            [0, 1, 0, 0, 0]], dtype=np.int32)
+            [0, 1, 0, 0, 0]], dtype=int32)
     >>> ll = [(0,2), (1,3), (4,)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[20,  5,  0],
             [ 5,  4,  1],
-            [ 0,  1,  0]], dtype=np.int32)
+            [ 0,  1,  0]], dtype=int32)
 
     Test removal of row/columns when the scaffold list does
     not contains all indices
     >>> ll = [(1,2), (3,)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[10,  2],
-            [ 2,  1]], dtype=np.int32)
+            [ 2,  1]], dtype=int32)
 
     Test with float and nan
     >>> A = csr_matrix(np.array([[0.1,0.1,0.2,0.2,np.nan],
     ... [0,0.1,0.2,0.2,1.1],
     ... [0,0,0.2,0.2,0], [0,0,0,0.1,0], [0,0,0,0,0]]), dtype=np.int32)
-    >>> dia = dia_matrix(([A.diagonal()], [0]), shape=A.shape)
-    >>> A= csr_matrix(A + A.T - dia)
+    >>> dia = dia_matrix(([A.diagonal()], [0]), shape=A.shape, dtype=np.int32)
+    >>> A= csr_matrix(A + A.T - dia, dtype=np.int32)
     >>> A.todense()
     matrix([[ 0.1,  0.1,  0.2,  0.2,  nan],
             [ 0.1,  0.1,  0.2,  0.2,  1.1],
             [ 0.2,  0.2,  0.2,  0.2,  0. ],
             [ 0.2,  0.2,  0.2,  0.1,  0. ],
-            [ nan,  1.1,  0. ,  0. ,  0. ]], dtype=np.int32)
+            [ nan,  1.1,  0. ,  0. ,  0. ]], dtype=int32)
 
     >>> ll = [(0,1), (2,3), (4,)]
     >>> reduce_matrix(A, ll, diagonal=True, use_triu=False).todense()
     matrix([[ 0.4,  0.8,  nan],
             [ 0.8,  0.7,  0. ],
-            [ nan,  0. ,  0. ]], dtype=np.int32)
+            [ nan,  0. ,  0. ]], dtype=int32)
     """
 
     if use_triu:
